cambios proyecto 5-1-2025

+++ my-ticketing-system/src/firebaseConfig.js
@@ -1,14 +1,14 @@
 import { initializeApp } from "firebase/app";
 import { getFirestore, collection, addDoc, setDoc, doc, getDocs, updateDoc, deleteDoc, getDoc } from "firebase/firestore";
+import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword } from "firebase/auth";
 import { getAnalytics } from "firebase/analytics";
 import { getStorage } from "firebase/storage";
 
-//configuración de Firebase
 const firebaseConfig = {
   apiKey: "AIzaSyD3SntYBXMfTyomUvfIeunYA9341F081Ok",
   authDomain: "proyectoindustrialtaurel.firebaseapp.com",
   projectId: "proyectoindustrialtaurel",
-  storageBucket: "proyectoindustrialtaurel.firebasestorage.app",
+  storageBucket: "proyectoindustrialtaurel.appspot.com",
   messagingSenderId: "797021083723",
   appId: "1:797021083723:web:60766602838b30773855ce",
   measurementId: "G-H843JF99Z1"
@@ -16,9 +16,8 @@ const firebaseConfig = {
 
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
+const auth = getAuth(app);
 const analytics = getAnalytics(app);
+const storage = getStorage(app);
 
-export const storage = getStorage(app);
-export { db, collection, addDoc, setDoc, doc, getDocs, updateDoc, deleteDoc, getDoc };
-
-
+export { db, collection, addDoc, setDoc, doc, getDocs, updateDoc, deleteDoc, getDoc, auth, createUserWithEmailAndPassword, signInWithEmailAndPassword };
